refactor(routing): remove dead ngrx router-store routing slice
The signal-native RoutingHistoryService replaced every reader of the
router-store routing state, so the whole chain is now dead:

- delete routingReducer + the `routing` AppState slice + its registration
- delete getPreviousRoutingState / getCurrentRoutingState selectors and
  the RoutingHistory / defaultRoutingState / AppRoutingOnlyAppState types
- remove StoreRouterConnectingModule + CustomRouterStateSerializer from
  AppModule (their ROUTER_NAVIGATION action only fed routingReducer)
- relocate RouterRedirect from routing.reducer.ts to router.actions.ts
  (still used by RouterNav + the auth slice) and repoint importers
- drop the now-pointless ROUTER_NAVIGATION dispatch + routing fixture
  from the table-cell-app-name spec and store test helper

RouterNav / RouterActions / RouterEffect stay: ~25 components plus the
auth flow still dispatch them for navigation. No @ngrx/router-store
imports remain in src.
