+++ src/frontend/packages/store/src/types/routing.type.ts
@@ -1,5 +1,3 @@
-import { AppRoutingOnlyAppState } from '../app-state';
-
 export interface RoutingEvent {
   id: number;
   url: string;
@@ -13,20 +11,4 @@ export interface RoutingEvent {
       [key: string]: string;
     }
   };
-}
-export interface RoutingHistory {
-  history?: RoutingEvent[];
-  currentState: RoutingEvent;
-  previousState?: RoutingEvent;
-}
-
-export const defaultRoutingState: RoutingHistory = {
-  currentState: null
-};
-
-export function getPreviousRoutingState(state: AppRoutingOnlyAppState) {
-  return state.routing.previousState;
-}
-export function getCurrentRoutingState(state: AppRoutingOnlyAppState) {
-  return state.routing.currentState;
 }
